requirements.txt
discord.py==1.4.1
requests===2.24.0
